requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
click==8.1.3
    # via -r requirements.in
numpy==1.21.2
    # via
    #   -r requirements.in
    #   pandas
pandas==1.4.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2022.1
    # via pandas
pyyaml==6.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
